Change number of packs in repo to decimal
Source: https://github.com/msupply-foundation/open-msupply/commit/05f500f6e2aa07582ef08ed372252d83ea58f8de

+++ server/repository/src/db_diesel/invoice_line.rs
@@ -58,7 +58,7 @@ pub struct InvoiceLineFilter {
     pub r#type: Option<EqualFilter<InvoiceLineRowType>>,
     pub location_id: Option<EqualFilter<String>>,
     pub requisition_id: Option<EqualFilter<String>>,
-    pub number_of_packs: Option<EqualFilter<i32>>,
+    pub number_of_packs: Option<EqualFilter<f64>>,
 }
 
 impl InvoiceLineFilter {
@@ -104,7 +104,7 @@ impl InvoiceLineFilter {
         self
     }
 
-    pub fn number_of_packs(mut self, filter: EqualFilter<i32>) -> Self {
+    pub fn number_of_packs(mut self, filter: EqualFilter<f64>) -> Self {
         self.number_of_packs = Some(filter);
         self
     }
